friedjof/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
asyncpg==0.30.0
attrs==25.3.0
croniter>=1.4.0
discord.py==2.5.2
frozenlist==1.6.2
greenlet==3.2.3
idna==3.10
loguru==0.7.3
multidict==6.4.4
propcache==0.3.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2024.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
yarl==1.20.0
